Подправил код

+++ all-code/algorithms 2.0/work3.cc
@@ -8,5 +8,5 @@ size_t CountInRange(const std::vector<int> &sorted_prices, int low, int high) {
     auto it1 = std::lower_bound(sorted_prices.begin(), sorted_prices.end(), low);
     auto it2 = std::upper_bound(sorted_prices.begin(), sorted_prices.end(), high);
 
-    return std::distance(it1, it2);
+    return static_cast<size_t>(std::distance(it1, it2));
 }
